+++ __tests__/services/grpc/world_engine_client_grpc.service.spec.ts
@@ -14,6 +14,7 @@ import { Config } from '../../../src/common/config';
 import { CLIENT_ID } from '../../../src/common/constants';
 import { Logger } from '../../../src/common/logger';
 import { PreviousDialog } from '../../../src/entities/continuation/previous_dialog.entity';
+import { SessionContinuation } from '../../../src/entities/continuation/session_continuation.entity';
 import { WorldEngineClientGrpcService } from '../../../src/services/gprc/world_engine_client_grpc.service';
 import { ExtendedInworldPacket } from '../../data_structures';
 import {
@@ -22,7 +23,9 @@ import {
   extendedCapabilities,
   extension,
   KEY,
+  phrases,
   previousDialog,
+  previousState,
   SCENE,
   SECRET,
   sessionContinuation,
@@ -42,27 +45,27 @@ describe('credentials', () => {
   });
 
   test('should use insecure credentials', () => {
-    const generateSessionToken = jest
-      .spyOn(Config.prototype, 'getSsl')
+    const getSSL = jest
+      .spyOn(Config.prototype, 'getSSL')
       .mockImplementationOnce(() => false);
 
     new WorldEngineClientGrpcService();
 
     expect(createSsl).toHaveBeenCalledTimes(0);
     expect(createInsecure).toHaveBeenCalledTimes(1);
-    expect(generateSessionToken).toHaveBeenCalledTimes(1);
+    expect(getSSL).toHaveBeenCalledTimes(1);
   });
 
   test('should use secure credentials', () => {
-    const generateSessionToken = jest
-      .spyOn(Config.prototype, 'getSsl')
+    const getSSL = jest
+      .spyOn(Config.prototype, 'getSSL')
       .mockImplementationOnce(() => true);
 
     new WorldEngineClientGrpcService();
 
     expect(createSsl).toHaveBeenCalledTimes(1);
     expect(createInsecure).toHaveBeenCalledTimes(0);
-    expect(generateSessionToken).toHaveBeenCalledTimes(1);
+    expect(getSSL).toHaveBeenCalledTimes(1);
   });
 });
 
@@ -185,7 +188,7 @@ describe('load scene', () => {
       name: SCENE,
       capabilities,
       sessionToken,
-      sessionContinuation: { previousDialog },
+      sessionContinuation: new SessionContinuation({ previousDialog: phrases }),
       user,
     });
 
@@ -236,6 +239,30 @@ describe('load scene', () => {
     expect(profileFields[0].getFieldValue()).toEqual(
       user.profile.fields[0].value,
     );
+  });
+
+  test('should send previous session state', async () => {
+    jest
+      .spyOn(WorldEngineClient.prototype, 'loadScene')
+      .mockImplementationOnce(mockLoadScene);
+
+    const sessionContinuation = new SessionContinuation({
+      previousState,
+    });
+
+    await client.loadScene({
+      name: SCENE,
+      capabilities,
+      sessionToken,
+      sessionContinuation,
+      user,
+    });
+
+    const state = mockLoadScene.mock.calls[0][0]
+      .getSessionContinuation()
+      .getPreviousState();
+
+    expect(state).toEqual(sessionContinuation.previousState);
   });
 
   test('should call extention functions', async () => {